Added the algorithm to plot magnetisation
The Metropolis algorithm now iterates over the lattice in order to bring it to equilibrium. Then it iterates again in order to measure the average value of a physical property of the system, here the magnetisation. The program does not yet yield the correct value for the critical temperature, but further work will be done tomorrow.