Added new file to compute magnetisation
The program from the magnetisation branch was pushed to the master branch as a new file. This program can be used to calculate magnetisation and will be update to calculate several other properties of the system.

The original program can be used to plot the Ising model's evolution over time for a given value of T.